requirements.txt
pygame-ce=2.5.1
noise=1.2.2
